I now store the ID in local storage and I also moved the display name to local storage
